requirements.txt
# Core LangChain dependencies (version 1.x)
langchain==1.0.2
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.1
langchain-openai==1.0.1
langchain-text-splitters==1.0.0

# LangGraph for agent workflows
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9

# LangSmith for tracing and monitoring
langsmith==0.4.38

# Document processing
pypdf==6.1.3
beautifulsoup4==4.14.2
bs4==0.0.2

# Cryptography
cryptography==46.0.3

# Environment variables
python-dotenv==1.2.1

# OpenAI
openai==1.109.1
tiktoken==0.12.0

# HTTP clients
httpx==0.28.1
httpx-sse==0.4.3
requests==2.32.5
requests-toolbelt==1.0.0

# Data processing
numpy==2.3.4
pandas==2.3.3
pyarrow==22.0.0

# Pydantic for data validation
pydantic==2.12.3
pydantic_core==2.41.4
pydantic-settings==2.11.0

# Database
SQLAlchemy==2.0.44

# Utilities
tenacity==9.1.2
tqdm==4.67.1
PyYAML==6.0.3
click==8.3.0
rich==14.2.0
colorama==0.4.6

# Jupyter (for development and testing)
ipython==9.6.0
ipykernel==7.1.0
jupyter_client==8.6.3
jupyter_core==5.9.1

# Evaluation and testing
ragas==0.3.7
datasets==4.3.0
scipy==1.16.2

# Code quality (optional, for development)
pre_commit==4.3.0
